feat(stage_l1_combined_syncer): Combined L1 syncing stages
