+++ eth/backend.go
@@ -806,7 +806,7 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 				// will trigger the staged sync which will require headers and blocks to be available
 				// in their respective cache in the download stage. If not found, it would cause a
 				// liveness issue for the chain.
-				if err := backend.sentriesClient.Hd.AddMinedHeader(b.Header()); err != nil {
+				if err = backend.sentriesClient.Hd.AddMinedHeader(b.Header()); err != nil {
 					logger.Error("add mined block to header downloader", "err", err)
 				}
 				backend.sentriesClient.Bd.AddToPrefetch(b.Header(), b.RawBody())
@@ -814,7 +814,7 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 				//p2p
 				//backend.sentriesClient.BroadcastNewBlock(context.Background(), b, b.Difficulty())
 				//rpcdaemon
-				if err := miningRPC.(*privateapi.MiningServer).BroadcastMinedBlock(b); err != nil {
+				if err = miningRPC.(*privateapi.MiningServer).BroadcastMinedBlock(b); err != nil {
 					logger.Error("txpool rpc mined block broadcast", "err", err)
 				}
 				logger.Trace("BroadcastMinedBlock successful", "number", b.Number(), "GasUsed", b.GasUsed(), "txn count", b.Transactions().Len())
@@ -1062,62 +1062,12 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 			log.Info("Starting sequencer in L1 recovery mode", "startBlock", cfg.L1SyncStartBlock)
 		}
 
-		seqAndVerifTopics := [][]libcommon.Hash{{
-			contracts.SequenceBatchesTopicPreEtrog,
-			contracts.SequenceBatchesTopicEtrog,
-			contracts.RollbackBatchesTopic,
-			contracts.VerificationTopicPreEtrog,
-			contracts.VerificationTopicEtrog,
-			contracts.VerificationValidiumTopicEtrog,
-		}}
-
-		seqAndVerifL1Contracts := []libcommon.Address{cfg.AddressRollup, cfg.AddressAdmin, cfg.AddressZkevm}
-
-		var l1Topics [][]libcommon.Hash
-		var l1Contracts []libcommon.Address
-		if isSequencer {
-			l1Topics = [][]libcommon.Hash{{
-				contracts.InitialSequenceBatchesTopic,
-				contracts.AddNewRollupTypeTopic,
-				contracts.AddNewRollupTypeTopicBanana,
-				contracts.CreateNewRollupTopic,
-				contracts.UpdateRollupTopic,
-			}}
-			l1Contracts = []libcommon.Address{cfg.AddressZkevm, cfg.AddressRollup}
-		} else {
-			l1Topics = seqAndVerifTopics
-			l1Contracts = seqAndVerifL1Contracts
-		}
-
 		ethermanClients := make([]syncer.IEtherman, len(backend.etherManClients))
 		for i, c := range backend.etherManClients {
 			ethermanClients[i] = c.EthClient
 		}
 
-		seqVerSyncer := syncer.NewL1Syncer(
-			ctx,
-			ethermanClients,
-			seqAndVerifL1Contracts,
-			seqAndVerifTopics,
-			cfg.L1BlockRange,
-			cfg.L1QueryDelay,
-			cfg.L1HighestBlockType,
-		)
-
-		backend.l1Syncer = syncer.NewL1Syncer(
-			ctx,
-			ethermanClients,
-			l1Contracts,
-			l1Topics,
-			cfg.L1BlockRange,
-			cfg.L1QueryDelay,
-			cfg.L1HighestBlockType,
-		)
-
 		log.Info("Rollup ID", "rollupId", cfg.L1RollupId)
-
-		// Check if L1 contracts addresses should be retrieved from the L1 chain
-		l1ContractAddressProcess(ctx, cfg.Zk, backend.l1Syncer)
 
 		l1InfoTreeSyncer := syncer.NewL1Syncer(
 			ctx,
@@ -1137,6 +1087,43 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 		}
 
 		if isSequencer {
+			l1Contracts := []libcommon.Address{
+				cfg.AddressZkevm,
+				cfg.AddressRollup,
+				cfg.AddressGerManager,
+			}
+			l1Topics := [][]libcommon.Hash{{
+				contracts.InitialSequenceBatchesTopic,
+				contracts.AddNewRollupTypeTopic,
+				contracts.AddNewRollupTypeTopicBanana,
+				contracts.CreateNewRollupTopic,
+				contracts.UpdateRollupTopic,
+				contracts.UpdateL1InfoTreeTopic,
+			}}
+
+			seqVerSyncer := syncer.NewL1Syncer(
+				ctx,
+				ethermanClients,
+				l1Contracts,
+				l1Topics,
+				cfg.L1BlockRange,
+				cfg.L1QueryDelay,
+				cfg.L1HighestBlockType,
+			)
+
+			backend.l1Syncer = syncer.NewL1Syncer(
+				ctx,
+				ethermanClients,
+				l1Contracts,
+				l1Topics,
+				cfg.L1BlockRange,
+				cfg.L1QueryDelay,
+				cfg.L1HighestBlockType,
+			)
+
+			// Check if L1 contracts addresses should be retrieved from the L1 chain
+			l1ContractAddressProcess(ctx, cfg.Zk, backend.l1Syncer)
+
 			// if we are sequencing transactions, we do the sequencing loop...
 			witnessGenerator := witness.NewGenerator(
 				config.Dirs,
@@ -1150,7 +1137,7 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 				backend.config.WitnessUnwindLimit,
 			)
 
-			var legacyExecutors []*legacy_executor_verifier.Executor = make([]*legacy_executor_verifier.Executor, 0, len(cfg.ExecutorUrls))
+			var legacyExecutors = make([]*legacy_executor_verifier.Executor, 0, len(cfg.ExecutorUrls))
 			if len(cfg.ExecutorUrls) > 0 && cfg.ExecutorUrls[0] != "" {
 				levCfg := legacy_executor_verifier.Config{
 					GrpcUrls:              cfg.ExecutorUrls,
@@ -1228,6 +1215,35 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 			  ZZZZZZZ  K   K  R   R   P       CCCC
 
 			*/
+
+			l1Contracts := []libcommon.Address{
+				cfg.AddressRollup,
+				cfg.AddressAdmin,
+				cfg.AddressZkevm,
+			}
+			l1Topics := [][]libcommon.Hash{{
+				contracts.SequenceBatchesTopicPreEtrog,
+				contracts.SequenceBatchesTopicEtrog,
+				contracts.RollbackBatchesTopic,
+				contracts.VerificationTopicPreEtrog,
+				contracts.VerificationTopicEtrog,
+				contracts.VerificationValidiumTopicEtrog,
+				contracts.UpdateL1InfoTreeTopic,
+			}}
+
+			backend.l1Syncer = syncer.NewL1Syncer(
+				ctx,
+				ethermanClients,
+				l1Contracts,
+				l1Topics,
+				cfg.L1BlockRange,
+				cfg.L1QueryDelay,
+				cfg.L1HighestBlockType,
+			)
+
+			// Check if L1 contracts addresses should be retrieved from the L1 chain
+			l1ContractAddressProcess(ctx, cfg.Zk, backend.l1Syncer)
+
 			latestForkId, err := stages.GetStageProgress(tx, stages.ForkId)
 			if err != nil {
 				return nil, err
